requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
openpyxl>=3.1.0
plotly>=5.14.0
gprofiler-official>=1.0.0
matplotlib>=3.7.0
pytest>=7.4.0
pytest-cov>=4.1.0
